requirements.txt
diffsort==0.2.0
hydra-core==1.3.2
hyperopt==0.2.7
ml-collections==0.1.1
numpy==1.26.4
protobuf==3.20.*
ray[tune]==2.30.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tqdm==4.66.5
torch==2.3.1
torchvision==0.18.1
